fix(month picker): format views and options #4190
