+++ projects/igniteui-angular/src/lib/calendar/calendar.component.html
@@ -45,6 +45,7 @@
                    [viewDate]="viewDate"
                    [weekStart]="weekStart"
                    [formatOptions]="formatOptions"
+                   [formatViews]="formatViews"
                    [selection]="selection"
                    [disabledDates]="disabledDates"
                    [specialDates]="specialDates"
@@ -53,10 +54,18 @@
     </igx-days-view>
 </div>
 
-<igx-months-view *ngIf="isYearView" [@animateView]="activeView" #months [date]="viewDate" [locale]="locale"
-    [monthFormat]="formatOptions.month" (onSelection)="changeMonth($event)">
+<igx-months-view *ngIf="isYearView" [@animateView]="activeView" #months 
+                 [date]="viewDate" 
+                 [locale]="locale"
+                 [formatView]="formatViews.month"
+                 [monthFormat]="formatOptions.month" 
+                 (onSelection)="changeMonth($event)">
 </igx-months-view>
 
-<igx-years-view *ngIf="isDecadeView" [@animateView]="activeView" #decade [date]="viewDate" [locale]="locale"
-    [yearFormat]="formatOptions.year" (onSelection)="changeYear($event)">
+<igx-years-view *ngIf="isDecadeView" [@animateView]="activeView" #decade 
+                [date]="viewDate" 
+                [locale]="locale"
+                [formatView]="formatViews.year"
+                [yearFormat]="formatOptions.year" 
+                (onSelection)="changeYear($event)">
 </igx-years-view>
