+++ src/app/calendar-views/calendar-views.sample.html
@@ -13,6 +13,8 @@
                                                   [viewDate]="viewDate"
                                                   [weekStart]="'1'"
                                                   [locale]="locale"
+                                                  [formatViews]="formatViews"
+                                                  [formatOptions]="formatOptions"
                                                   [specialDates]="specialDates"
                                                   [disabledDates]="disabledDates"
                                                   [selection]="'multi'"
